Program 12
Still needs the win conditions finished, but those are stupid and I hate
them.

The board needs to be created, but fancy ASCII connect-four boards can
wait.

Still need to figure out a way to test a random move for legality
without calling (random) twice. Coming up with actual AI would solve
this, but we'll see if I have time.